tests/requirements.txt
pytest==8.3.4
requests==2.32.3
psycopg[binary]==3.2.6
PyYAML==6.0.2
